extract `svg` from `Schema` block to separate file
