+++ src/components/index/SchemaSVG.tsx
@@ -1,0 +1,184 @@
+interface SchemaSVGProps {
+  platform: string
+  onMouseEnter?: (id: string) => void
+  onMouseLeave?: () => void
+}
+
+export default function SchemaSVG({platform, onMouseEnter, onMouseLeave}: SchemaSVGProps) {
+  const isDesktop = platform === 'desktop'
+
+  return (
+    <>
+      <svg className="w-full" viewBox="0 0 821 646" fill="none" xmlns="http://www.w3.org/2000/svg">
+        <g id="fields" {...(isDesktop && {onMouseEnter: () => onMouseEnter && onMouseEnter('fields'), onMouseLeave})}>
+          <path fill="#fff" d="M430.5 2h389v642h-389z" />
+          <path stroke="#454545" strokeWidth="3" d="M430.5 2h389v642h-389z" />
+          <g id="Frame 199">
+            <mask id="path-2-inside-1_3482_3185" fill="#fff">
+              <rect width="352" height="175" x="449" y="30.5" rx="2" />
+            </mask>
+            <rect width="352" height="175" x="449" y="30.5" fill="#fff" rx="2" />
+            <rect width="352" height="175" x="449" y="30.5" stroke="#454545" strokeWidth="6" mask="url(#path-2-inside-1_3482_3185)" rx="2" />
+            <path id="Rectangle 85" fill="#454545" d="M624 30.5h3v175h-3z" />
+            <circle id="Ellipse 1" cx="625" cy="118.5" r="48.5" stroke="#454545" strokeWidth="3" />
+            <path id="Ellipse 2" stroke="#454545" strokeWidth="3" d="M450.5 70.023a48.498 48.498 0 0 1 47 48.477 48.512 48.512 0 0 1-14.205 34.295 48.512 48.512 0 0 1-32.795 14.182V70.023Z" />
+            <path id="Ellipse 3" stroke="#454545" strokeWidth="3" d="M799.5 166.977a48.512 48.512 0 0 1-32.795-14.182A48.512 48.512 0 0 1 752.5 118.5a48.487 48.487 0 0 1 14.205-34.295A48.51 48.51 0 0 1 799.5 70.023v96.954Z" />
+          </g>
+          <g id="Frame 200">
+            <mask id="path-7-inside-2_3482_3185" fill="#fff">
+              <rect width="352" height="175" x="449" y="235.5" rx="2" />
+            </mask>
+            <rect width="352" height="175" x="449" y="235.5" fill="#fff" rx="2" />
+            <rect width="352" height="175" x="449" y="235.5" stroke="#454545" strokeWidth="6" mask="url(#path-7-inside-2_3482_3185)" rx="2" />
+            <path id="Rectangle 85_2" fill="#454545" d="M624 235.5h3v175h-3z" />
+            <circle id="Ellipse 1_2" cx="625" cy="323.5" r="48.5" stroke="#454545" strokeWidth="3" />
+            <path id="Ellipse 2_2" stroke="#454545" strokeWidth="3" d="M450.5 275.023a48.512 48.512 0 0 1 32.795 14.182A48.512 48.512 0 0 1 497.5 323.5a48.512 48.512 0 0 1-14.205 34.295 48.512 48.512 0 0 1-32.795 14.182v-96.954Z" />
+            <path id="Ellipse 3_2" stroke="#454545" strokeWidth="3" d="M799.5 371.977a48.512 48.512 0 0 1-32.795-14.182A48.512 48.512 0 0 1 752.5 323.5a48.512 48.512 0 0 1 14.205-34.295 48.512 48.512 0 0 1 32.795-14.182v96.954Z" />
+          </g>
+          <g id="Frame 201">
+            <mask id="path-12-inside-3_3482_3185" fill="#fff">
+              <rect width="352" height="175" x="449" y="440.5" rx="2" />
+            </mask>
+            <rect width="352" height="175" x="449" y="440.5" fill="#fff" rx="2" />
+            <rect width="352" height="175" x="449" y="440.5" stroke="#454545" strokeWidth="6" mask="url(#path-12-inside-3_3482_3185)" rx="2" />
+            <path id="Rectangle 85_3" fill="#454545" d="M624 440.5h3v175h-3z" />
+            <circle id="Ellipse 1_3" cx="625" cy="528.5" r="48.5" stroke="#454545" strokeWidth="3" />
+            <path id="Ellipse 2_3" stroke="#454545" strokeWidth="3" d="M450.5 480.023a48.512 48.512 0 0 1 32.795 14.182A48.512 48.512 0 0 1 497.5 528.5a48.512 48.512 0 0 1-14.205 34.295 48.512 48.512 0 0 1-32.795 14.182v-96.954Z" />
+            <path id="Ellipse 3_3" stroke="#454545" strokeWidth="3" d="M799.5 576.977a48.512 48.512 0 0 1-32.795-14.182A48.512 48.512 0 0 1 752.5 528.5a48.512 48.512 0 0 1 14.205-34.295 48.512 48.512 0 0 1 32.795-14.182v96.954Z" />
+          </g>
+        </g>
+        <g id="parking" {...(isDesktop && {onMouseEnter: () => onMouseEnter && onMouseEnter('parking'), onMouseLeave})}>
+          <path fill="#fff" d="M0 381.5h340v201H0z" />
+          <g id="Union" fill="#454545">
+            <path fillRule="evenodd" d="m81.757 383.102-.137-1.156c1.025-.107 1.788-.195 2.383-.264 1.055-.121 1.583-.182 2.11-.182.53 0 1.059.061 2.118.182.596.069 1.36.157 2.388.264l-.139 1.156 2.24.998-.35-1.009-1.57-.53.052-.541 1.719.52 1.22 2.345.022 7.609.35.072c1.449.299 2.259.465 1.973 1.509l-2.323-.509.074 21.637-.626 1.846-.87.403c-4.414.064-8.14.064-12.544 0l-.87-.403-.626-1.846.064-21.648-2.356.52c-.29-1.039.519-1.207 1.95-1.505l.416-.087.021-7.598 1.21-2.345 1.72-.52.063.541-1.581.53-.35 1.009 2.25-.998Zm-1.793 30.212 2.367.573a111.27 111.27 0 0 1 7.576 0l2.367-.573-1.486-3.789c-3.12.011-6.218.011-9.338 0l-1.486 3.789Zm-.393-19.537 1.168 3.598c1.316-.626 9.444-.626 10.76 0l1.167-3.598a69.37 69.37 0 0 1-.814-.36v-.001c-1.602-.717-2.78-1.244-4.8-1.134h-1.867c-2.018-.11-3.195.417-4.797 1.134-.259.116-.53.237-.817.361Zm12.342 4.351.892-2.292v11.429l-1.125.594.233-9.731Zm-11.6 0-.89-2.292.01 11.429 1.115.594-.234-9.731Z" clipRule="evenodd" />
+            <path d="M39 381.5h-3v42H0v3h36v40h3v-40h30v40h3v-40h30v40h3v-40h30v40h3v-40h30v40h3v-40h30v40h3v-40h30v40h3v-40h30v40h3v-40h30v40h3v-40h37v-3h-37v-42h-3v42h-30v-42h-3v42h-30v-42h-3v42h-30v-42h-3v42h-30v-42h-3v42h-30v-42h-3v42h-30v-42h-3v42H72v-42h-3v42H39v-42Z" />
+            <path
+              fillRule="evenodd"
+              d="m115.063 381.946.138 1.156-2.251.998.351-1.009 1.581-.53-.063-.541-1.719.52-1.211 2.345-.021 7.598-.416.087c-1.431.298-2.241.466-1.95 1.505l2.356-.52-.064 21.648.627 1.846.87.403c4.403.064 8.129.064 12.543 0l.87-.403.626-1.846-.074-21.637 2.324.509c.285-1.044-.525-1.21-1.974-1.509l-.35-.072-.022-7.609-1.22-2.345-1.719-.52-.053.541 1.57.53.351 1.009-2.239-.998.138-1.156a196.981 196.981 0 0 1-2.388-.264c-1.059-.121-1.588-.182-2.117-.182-.528 0-1.056.061-2.111.182-.595.069-1.358.157-2.383.264Zm.711 31.941-2.367-.573 1.486-3.789c3.12.011 6.219.011 9.338 0l1.486 3.789-2.367.573a111.264 111.264 0 0 0-7.576 0Zm-1.592-16.512-1.167-3.598c.287-.124.557-.245.817-.361 1.601-.717 2.778-1.244 4.797-1.134h1.867c2.019-.11 3.197.417 4.799 1.134l.001.001c.258.116.528.236.814.36l-1.168 3.598c-1.316-.626-9.444-.626-10.76 0Zm12.066-1.539-.892 2.292-.233 9.731 1.125-.594v-11.429Zm-13.382 0 .891 2.292.234 9.731-1.114-.594-.011-11.429ZM148.644 383.102l-.138-1.156c1.025-.107 1.788-.195 2.383-.264 1.055-.121 1.583-.182 2.111-.182.529 0 1.058.061 2.117.182.596.069 1.361.157 2.388.264l-.138 1.156 2.239.998-.35-1.009-1.571-.53.053-.541 1.719.52 1.22 2.345.022 7.609.35.072c1.449.299 2.259.465 1.974 1.509l-2.324-.509.074 21.637-.625 1.846-.871.403a409.73 409.73 0 0 1-12.543 0l-.87-.403-.627-1.846.064-21.648-2.356.52c-.29-1.039.519-1.207 1.95-1.505h.003l.413-.087.022-7.598 1.21-2.345 1.719-.52.063.541-1.581.53-.35 1.009 2.25-.998Zm-1.794 30.212 2.367.573c2.918-.086 4.648-.086 7.576 0l2.367-.573-1.486-3.789c-3.119.011-6.218.011-9.338 0l-1.486 3.789Zm-.392-19.537 1.167 3.598c1.316-.626 9.445-.626 10.761 0l1.167-3.598a65.485 65.485 0 0 1-.813-.36l-.002-.001c-1.602-.717-2.779-1.244-4.799-1.134h-1.867c-2.019-.11-3.196.417-4.797 1.134-.259.116-.53.237-.817.361Zm12.342 4.351.891-2.292v11.429l-1.124.594.233-9.731Zm-11.6 0-.891-2.292.011 11.429 1.114.594-.234-9.731ZM181.949 381.946l.138 1.156-2.25.998.35-1.009 1.581-.53-.063-.541-1.719.52-1.21 2.345-.022 7.598-.413.087h-.002c-1.431.298-2.241.466-1.951 1.505l2.356-.52-.063 21.648.626 1.846.87.403c4.404.064 8.129.064 12.543 0l.871-.403.625-1.846-.073-21.637 2.323.509c.285-1.044-.525-1.21-1.973-1.509l-.35-.072-.022-7.609-1.221-2.345-1.719-.52-.052.541 1.57.53.35 1.009-2.239-.998.138-1.156a199.281 199.281 0 0 1-2.388-.264c-1.059-.121-1.588-.182-2.117-.182-.528 0-1.056.061-2.11.182-.596.069-1.359.157-2.384.264Zm.711 31.941-2.367-.573 1.486-3.789c3.121.011 6.219.011 9.339 0l1.486 3.789-2.367.573a111.294 111.294 0 0 0-7.577 0Zm-1.592-16.512-1.167-3.598c.287-.124.558-.245.817-.361 1.601-.717 2.779-1.244 4.797-1.134h1.868c2.019-.11 3.196.417 4.798 1.134l.001.001c.258.115.528.236.814.36l-1.167 3.598c-1.316-.626-9.445-.626-10.761 0Zm12.066-1.539-.891 2.292-.233 9.731 1.124-.594v-11.429Zm-13.381 0 .89 2.292.234 9.731-1.114-.594-.01-11.429ZM81.757 432.102l-.137-1.156c1.025-.107 1.788-.195 2.383-.264 1.055-.121 1.583-.182 2.11-.182.53 0 1.058.061 2.117.182.597.069 1.362.157 2.389.264l-.139 1.156 2.24.998-.35-1.009-1.57-.53.052-.541 1.719.52 1.22 2.345.022 7.609.35.072c1.448.298 2.259.465 1.973 1.509l-2.323-.509.074 21.637-.626 1.846-.87.403c-4.414.064-8.14.064-12.544 0l-.87-.403-.626-1.846.064-21.648-2.356.52c-.29-1.039.519-1.207 1.95-1.505l.416-.087.021-7.598 1.21-2.345 1.72-.52.063.541-1.581.53-.35 1.009 2.25-.998Zm-1.793 30.212 2.367.573a111.27 111.27 0 0 1 7.576 0l2.367-.573-1.486-3.789c-3.12.011-6.218.011-9.338 0l-1.486 3.789Zm-.393-19.537 1.168 3.598c1.316-.626 9.444-.626 10.76 0l1.167-3.598c-.286-.124-.556-.245-.815-.361-1.601-.717-2.779-1.244-4.798-1.134h-1.868c-2.018-.11-3.195.417-4.797 1.134-.259.116-.53.237-.817.361Zm12.342 4.351.892-2.292v11.429l-1.125.594.233-9.731Zm-11.6 0-.89-2.292.01 11.429 1.115.594-.234-9.731ZM115.063 430.946l.138 1.156-2.251.998.351-1.009 1.581-.53-.063-.541-1.719.52-1.211 2.345-.021 7.598c-.144.031-.282.06-.416.087-1.431.298-2.241.466-1.95 1.505l2.356-.52-.064 21.648.627 1.846.87.403c4.403.064 8.129.064 12.543 0l.87-.403.626-1.846-.074-21.637 2.324.509c.285-1.044-.525-1.21-1.974-1.509l-.35-.072-.022-7.609-1.22-2.345-1.719-.52-.053.541 1.57.53.351 1.009-2.239-.998.138-1.156a199.281 199.281 0 0 1-2.388-.264c-1.059-.121-1.588-.182-2.117-.182-.528 0-1.056.061-2.111.182-.595.069-1.358.157-2.383.264Zm.711 31.941-2.367-.573 1.486-3.789c3.12.011 6.219.011 9.338 0l1.486 3.789-2.367.573a111.264 111.264 0 0 0-7.576 0Zm-1.592-16.512-1.167-3.598c.287-.124.557-.245.817-.361 1.601-.717 2.778-1.244 4.797-1.134h1.867c2.019-.11 3.197.417 4.799 1.134.258.116.528.237.815.361l-1.168 3.598c-1.316-.626-9.444-.626-10.76 0Zm12.066-1.539-.892 2.292-.233 9.731 1.125-.594v-11.429Zm-13.382 0 .891 2.292.234 9.731-1.114-.594-.011-11.429Z"
+              clipRule="evenodd"
+            />
+            <path d="M36 497.5h3v42h30v-42h3v42h30v-42h3v42h30v-42h3v42h30v-42h3v42h30v-42h3v42h30v-42h3v42h30v-42h3v42h30v-42h3v42h37v3h-37v40h-3v-40h-30v40h-3v-40h-30v40h-3v-40h-30v40h-3v-40h-30v40h-3v-40h-30v40h-3v-40h-30v40h-3v-40H72v40h-3v-40H39v40h-3v-40H0v-3h36v-42Z" />
+            <path
+              fillRule="evenodd"
+              d="m247.757 499.102-.137-1.156c1.025-.107 1.788-.195 2.383-.264 1.055-.121 1.583-.182 2.111-.182.529 0 1.058.061 2.117.182.596.069 1.36.157 2.388.264l-.139 1.156 2.239.998-.35-1.009-1.57-.53.053-.541 1.719.52 1.22 2.345.022 7.609.35.072c1.449.299 2.259.465 1.973 1.509l-2.323-.509.074 21.637-.626 1.846-.87.403c-4.414.064-8.14.064-12.543 0l-.87-.403-.627-1.846.064-21.648-2.356.52c-.291-1.039.519-1.207 1.95-1.505l.416-.087.021-7.598 1.211-2.345 1.719-.52.063.541-1.582.53-.35 1.009 2.25-.998Zm-1.793 30.212 2.367.573c2.917-.086 4.647-.086 7.576 0l2.367-.573-1.486-3.789c-3.119.011-6.218.011-9.338 0l-1.486 3.789Zm-.393-19.537 1.168 3.598c1.316-.626 9.444-.626 10.76 0l1.167-3.598a43.876 43.876 0 0 1-.347-.152l-.145-.064-.14-.063-.183-.082c-1.601-.717-2.779-1.244-4.798-1.134h-1.868c-2.018-.11-3.195.417-4.797 1.134-.259.116-.53.237-.817.361Zm12.342 4.351.892-2.292v11.429l-1.125.594.233-9.731Zm-11.6 0-.89-2.292.011 11.429 1.114.594-.235-9.731ZM281.063 497.946l.138 1.156-2.251.998.351-1.009 1.581-.53-.063-.541-1.719.52-1.211 2.345-.021 7.598-.416.087c-1.431.298-2.241.466-1.95 1.505l2.356-.52-.064 21.648.627 1.846.87.403c4.403.064 8.129.064 12.543 0l.87-.403.626-1.846-.074-21.637 2.324.509c.285-1.044-.525-1.21-1.974-1.509l-.35-.072-.022-7.609-1.22-2.345-1.719-.52-.053.541 1.57.53.351 1.009-2.24-.998.139-1.156a196.981 196.981 0 0 1-2.388-.264c-1.059-.121-1.588-.182-2.117-.182-.528 0-1.056.061-2.111.182-.595.069-1.358.157-2.383.264Zm.711 31.941-2.367-.573 1.486-3.789c3.12.011 6.219.011 9.338 0l1.486 3.789-2.367.573a111.264 111.264 0 0 0-7.576 0Zm-1.592-16.512-1.167-3.598c.286-.124.557-.245.816-.361 1.602-.717 2.779-1.244 4.798-1.134h1.867c2.019-.11 3.197.417 4.799 1.134l.001.001c.258.116.528.236.814.36l-1.168 3.598c-1.316-.626-9.444-.626-10.76 0Zm12.066-1.539-.892 2.292-.233 9.731 1.125-.594v-11.429Zm-13.382 0 .891 2.292.234 9.731-1.114-.594-.011-11.429ZM214.757 548.102l-.137-1.156c1.025-.107 1.788-.195 2.383-.264 1.054-.121 1.583-.182 2.111-.182.529 0 1.057.061 2.116.182.597.069 1.362.157 2.389.264l-.139 1.156 2.239.998-.35-1.009-1.57-.53.053-.541 1.719.52 1.22 2.345.022 7.609.131.027.219.045c1.449.299 2.259.465 1.973 1.509l-2.323-.509.074 21.637-.626 1.846-.87.403c-4.414.064-8.14.064-12.543 0l-.87-.403-.627-1.846.064-21.648-2.356.52c-.291-1.039.519-1.207 1.95-1.505l.218-.045.198-.042.021-7.598 1.211-2.345 1.719-.52.063.541-1.582.53-.35 1.009 2.25-.998Zm-1.793 30.212 2.367.573c2.917-.086 4.647-.086 7.576 0l2.367-.573-1.486-3.789c-3.119.011-6.218.011-9.338 0l-1.486 3.789Zm-.393-19.537 1.168 3.598c1.316-.626 9.444-.626 10.76 0l1.167-3.598c-.286-.124-.556-.245-.815-.361-1.601-.717-2.779-1.244-4.798-1.134h-1.868c-2.018-.11-3.195.417-4.797 1.134-.259.116-.53.237-.817.361Zm12.342 4.351.892-2.292v11.429l-1.125.594.233-9.731Zm-11.6 0-.89-2.292.011 11.429 1.114.594-.235-9.731ZM248.063 546.946l.138 1.156-2.251.998.351-1.009 1.581-.53-.063-.541-1.719.52-1.211 2.345-.021 7.598-.416.087c-1.431.298-2.241.466-1.95 1.505l2.356-.52-.064 21.648.627 1.846.87.403c4.403.064 8.129.064 12.543 0l.87-.403.626-1.846-.074-21.637 2.324.509c.285-1.044-.525-1.21-1.974-1.509l-.35-.072-.022-7.609-1.22-2.345-1.719-.52-.053.541 1.57.53.351 1.009-2.24-.998.139-1.156a196.981 196.981 0 0 1-2.388-.264c-1.059-.121-1.588-.182-2.117-.182-.528 0-1.056.061-2.11.182-.596.069-1.359.157-2.384.264Zm.711 31.941-2.367-.573 1.486-3.789c3.12.011 6.219.011 9.338 0l1.486 3.789-2.367.573a111.264 111.264 0 0 0-7.576 0Zm-1.592-16.512-1.167-3.598c.285-.124.555-.244.813-.36l.003-.001c1.602-.717 2.779-1.244 4.798-1.134h1.867c2.019-.11 3.197.417 4.799 1.134.258.116.528.237.815.361l-1.168 3.598c-1.316-.626-9.444-.626-10.76 0Zm12.066-1.539-.892 2.292-.233 9.731 1.125-.594v-11.429Zm-13.382 0 .891 2.292.234 9.731-1.114-.594-.011-11.429ZM115.757 548.102l-.137-1.156c1.025-.107 1.788-.195 2.383-.264 1.054-.121 1.583-.182 2.111-.182.529 0 1.058.061 2.117.182.596.069 1.361.157 2.388.264l-.139 1.156 2.239.998-.35-1.009-1.57-.53.053-.541 1.719.52 1.22 2.345.022 7.609.35.072c1.449.299 2.259.465 1.973 1.509l-2.323-.509.074 21.637-.626 1.846-.87.403c-4.414.064-8.14.064-12.543 0l-.87-.403-.627-1.846.064-21.648-2.356.52c-.291-1.039.519-1.207 1.95-1.505l.416-.087.021-7.598 1.211-2.345 1.719-.52.063.541-1.582.53-.35 1.009 2.25-.998Zm-1.793 30.212 2.367.573c2.917-.086 4.647-.086 7.576 0l2.367-.573-1.486-3.789c-3.119.011-6.218.011-9.338 0l-1.486 3.789Zm-.393-19.537 1.168 3.598c1.316-.626 9.444-.626 10.76 0l1.167-3.598c-.286-.124-.556-.245-.815-.361-1.601-.717-2.779-1.244-4.798-1.134h-1.868c-2.018-.11-3.195.417-4.797 1.134l-.002.001c-.259.115-.528.236-.815.36Zm12.342 4.351.892-2.292v11.429l-1.125.594.233-9.731Zm-11.6 0-.89-2.292.011 11.429 1.114.594-.235-9.731ZM48.62 546.946l.137 1.156-2.25.998.35-1.009 1.582-.53-.063-.541-1.72.52-1.21 2.345-.021 7.598-.416.087c-1.431.298-2.24.466-1.95 1.505l2.356-.52-.064 21.648.627 1.846.87.403c4.403.064 8.129.064 12.543 0l.87-.403.626-1.846-.074-21.637 2.323.509c.286-1.044-.524-1.21-1.973-1.509h-.002a91.862 91.862 0 0 1-.348-.072l-.022-7.609-1.22-2.345-1.72-.52-.052.541 1.57.53.35 1.009-2.239-.998.139-1.156a197.152 197.152 0 0 1-2.388-.264c-1.06-.121-1.588-.182-2.117-.182-.528 0-1.056.061-2.111.182-.595.069-1.358.157-2.383.264Zm.71 31.941-2.366-.573 1.486-3.789c3.12.011 6.219.011 9.338 0l1.486 3.789-2.367.573a111.27 111.27 0 0 0-7.576 0Zm-1.591-16.512-1.168-3.598c.287-.124.558-.245.817-.361 1.601-.717 2.779-1.244 4.797-1.134h1.868c2.02-.11 3.197.417 4.798 1.134.26.116.53.237.815.361l-1.167 3.598c-1.316-.626-9.444-.626-10.76 0Zm12.066-1.539-.892 2.292-.233 9.731 1.125-.594v-11.429Zm-13.382 0 .89 2.292.235 9.731-1.114-.594-.011-11.429Z"
+              clipRule="evenodd"
+            />
+          </g>
+        </g>
+        <g id="_">
+          <path fill="#fff" d="M217.5 2h188v265h-188z" />
+          <path stroke="#454545" strokeWidth="3" d="M217.5 2h188v265h-188z" />
+          <path stroke="#454545" strokeWidth="3" d="M348.5 210h57v57h-57z" />
+          <path stroke="#454545" strokeWidth="3" d="M355.5 160h25v25h-25z" />
+          <path stroke="#454545" strokeWidth="3" d="M405.5 185h25v25h-25z" />
+        </g>
+        <g id="gymnast" {...(isDesktop && {onMouseEnter: () => onMouseEnter && onMouseEnter('gymnast'), onMouseLeave})}>
+          <rect x="305.5" y="59.5" width="100" height="100" fill="white" />
+          <rect x="305.5" y="59.5" width="100" height="100" stroke="#454545" strokeWidth="3" />
+          <g id="Vector">
+            <path d="M364.778 97.1053C364.996 96.3061 365.821 95.8351 366.62 96.0534C367.372 96.2587 367.834 97.0014 367.703 97.7543L367.672 97.8955L366.395 102.57C366.373 102.653 366.343 102.734 366.307 102.811L366.247 102.925L362.747 108.987C362.701 109.067 362.649 109.142 362.591 109.21C362.525 109.303 362.451 109.391 362.368 109.474L358.271 113.571L359.332 115.41C359.438 115.592 359.503 115.794 359.525 116.002L359.533 116.16V122.204C359.533 123.032 358.862 123.704 358.033 123.704C357.254 123.704 356.613 123.109 356.54 122.348L356.533 122.204L356.533 116.561L354.245 112.598C354.227 112.572 354.209 112.545 354.191 112.517C353.781 111.884 353.633 111.129 353.749 110.409C353.846 109.808 354.126 109.231 354.59 108.767L355.966 107.391H350.549C349.72 107.391 349.049 106.72 349.049 105.891C349.049 105.112 349.644 104.471 350.404 104.398L350.549 104.391H361.937L363.552 101.593L364.778 97.1053Z" fill="#454545" />
+            <path d="M359.711 103.767C361.368 103.767 362.711 102.423 362.711 100.767C362.711 99.1098 361.368 97.7666 359.711 97.7666C358.054 97.7666 356.711 99.1098 356.711 100.767C356.711 102.423 358.054 103.767 359.711 103.767Z" fill="#454545" />
+            <path d="M352.952 109.517L350.655 109.517L344.876 108.022L344.734 107.993C343.98 107.873 343.243 108.344 343.048 109.099C342.841 109.901 343.323 110.719 344.125 110.927L350.091 112.47L350.214 112.496C350.298 112.51 350.382 112.517 350.466 112.517H353.055C352.633 111.568 352.598 110.488 352.952 109.517Z" fill="#454545" />
+          </g>
+        </g>
+        <g id="cloakroom" {...(isDesktop && {onMouseEnter: () => onMouseEnter && onMouseEnter('cloakroom'), onMouseLeave})}>
+          <rect x="380.5" y="160" width="25" height="25" fill="white" />
+          <path stroke="#454545" strokeWidth="3" d="M380.5 160h25v25h-25z" />
+          <path id="Vector" fill="#454545" d="M399.835 176.41c0 .465-.131.916-.374 1.306-.241.389-.6.721-1.051.923-.3.134-.64.208-.998.207h-8.824a2.414 2.414 0 0 1-.997-.208 2.36 2.36 0 0 1-.772-.554 2.474 2.474 0 0 1-.653-1.674 2.357 2.357 0 0 1 .218-.998 2.43 2.43 0 0 1 .677-.863l4.806-3.914-.226-.183a2.417 2.417 0 0 1-.669-2.894 2.416 2.416 0 0 1 3.724-.86l.711.58-.905 1.111-.712-.58a.982.982 0 0 0-1.608.767.97.97 0 0 0 .364.764l.907.739 5.487 4.47a2.372 2.372 0 0 1 .895 1.861Zm-1.505.376c.048-.12.072-.249.072-.376a.942.942 0 0 0-.367-.75L393 171.559l-5.034 4.102a.978.978 0 0 0-.282.354.927.927 0 0 0-.085.395c-.001.19.055.386.157.549a.948.948 0 0 0 .832.454h8.824c.123 0 .228-.018.323-.049a.92.92 0 0 0 .254-.129.987.987 0 0 0 .341-.449Z" />
+        </g>
+        <g id="toilets" {...(isDesktop && {onMouseEnter: () => onMouseEnter && onMouseEnter('toilets'), onMouseLeave})}>
+          <rect width="103" height="28" transform="translate(304 33.5)" fill="white" />
+          <g id="Frame 213">
+            <path stroke="#454545" strokeWidth="3" d="M305.5 35h25v25h-25z" />
+            <g id="Group">
+              <g id="Group_2">
+                <g id="Group_3" fill="#454545">
+                  <path id="Vector_2" d="M312.91 45.096h-.239a2.259 2.259 0 0 0-2.251 2.251v2.219a.64.64 0 0 0 .638.639h.527v2.777c0 .495.399.894.894.894h.607c.495 0 .894-.399.894-.894v-2.777h.527a.64.64 0 0 0 .638-.639v-2.219a2.225 2.225 0 0 0-2.235-2.25Z" />
+                  <path id="Vector_3" d="M312.799 44.634a1.756 1.756 0 1 0 0-3.512 1.756 1.756 0 0 0 0 3.512Z" />
+                  <path id="Vector_4" d="m325.553 50.715-1.165-4.406a1.62 1.62 0 0 0-1.564-1.213c-.735 0-1.389.495-1.565 1.213l-1.165 4.406a.735.735 0 0 0 .718.926h.83l.447 1.82a.562.562 0 0 0 .543.431h.415a.562.562 0 0 0 .543-.43l.431-1.82h.83c.479 0 .83-.464.702-.927Z" />
+                  <path id="Vector_5" d="M322.823 44.634a1.756 1.756 0 1 0 0-3.512 1.756 1.756 0 0 0 0 3.512Z" />
+                  <path id="Rectangle 99" d="M317.188 41.108h1.247v12.774h-1.247z" />
+                </g>
+              </g>
+            </g>
+          </g>
+          <g id="Frame 215">
+            <path stroke="#454545" strokeWidth="3" d="M330.5 35h25v25h-25z" />
+            <g id="Group_4">
+              <g id="Group_5">
+                <g id="Group_6" fill="#454545">
+                  <path id="Vector_6" d="M337.91 45.096h-.239a2.259 2.259 0 0 0-2.251 2.251v2.219a.64.64 0 0 0 .638.639h.527v2.777c0 .495.399.894.894.894h.607c.495 0 .894-.399.894-.894v-2.777h.527a.64.64 0 0 0 .638-.639v-2.219a2.225 2.225 0 0 0-2.235-2.25Z" />
+                  <path id="Vector_7" d="M337.799 44.634a1.756 1.756 0 1 0 0-3.512 1.756 1.756 0 0 0 0 3.512Z" />
+                  <path id="Vector_8" d="m350.553 50.715-1.165-4.406a1.62 1.62 0 0 0-1.564-1.213c-.735 0-1.389.495-1.565 1.213l-1.165 4.406a.735.735 0 0 0 .718.926h.83l.447 1.82a.562.562 0 0 0 .543.431h.415a.562.562 0 0 0 .543-.43l.431-1.82h.83c.479 0 .83-.464.702-.927Z" />
+                  <path id="Vector_9" d="M347.823 44.634a1.756 1.756 0 1 0 0-3.512 1.756 1.756 0 0 0 0 3.512Z" />
+                  <path id="Rectangle 99_2" d="M342.188 41.108h1.247v12.774h-1.247z" />
+                </g>
+              </g>
+            </g>
+          </g>
+          <g id="Frame 216">
+            <path stroke="#454545" strokeWidth="3" d="M355.5 35h25v25h-25z" />
+            <g id="Group_7">
+              <g id="Group_8">
+                <g id="Group_9" fill="#454545">
+                  <path id="Vector_10" d="M362.91 45.096h-.239a2.259 2.259 0 0 0-2.251 2.251v2.219a.64.64 0 0 0 .638.639h.527v2.777c0 .495.399.894.894.894h.607c.495 0 .894-.399.894-.894v-2.777h.527a.64.64 0 0 0 .638-.639v-2.219a2.225 2.225 0 0 0-2.235-2.25Z" />
+                  <path id="Vector_11" d="M362.799 44.634a1.756 1.756 0 1 0 0-3.512 1.756 1.756 0 0 0 0 3.512Z" />
+                  <path id="Vector_12" d="m375.553 50.715-1.165-4.406a1.62 1.62 0 0 0-1.564-1.213c-.735 0-1.389.495-1.565 1.213l-1.165 4.406a.735.735 0 0 0 .718.926h.83l.447 1.82a.562.562 0 0 0 .543.431h.415a.562.562 0 0 0 .543-.43l.431-1.82h.83c.479 0 .83-.464.702-.927Z" />
+                  <path id="Vector_13" d="M372.823 44.634a1.756 1.756 0 1 0 0-3.512 1.756 1.756 0 0 0 0 3.512Z" />
+                  <path id="Rectangle 99_3" d="M367.188 41.108h1.247v12.774h-1.247z" />
+                </g>
+              </g>
+            </g>
+          </g>
+          <g id="Frame 214">
+            <path stroke="#454545" strokeWidth="3" d="M380.5 35h25v25h-25z" />
+            <g id="Group_10">
+              <g id="Group_11">
+                <g id="Group_12" fill="#454545">
+                  <path id="Vector_14" d="M387.91 45.096h-.239a2.259 2.259 0 0 0-2.251 2.251v2.219a.64.64 0 0 0 .638.639h.527v2.777c0 .495.399.894.894.894h.607c.495 0 .894-.399.894-.894v-2.777h.527a.64.64 0 0 0 .638-.639v-2.219a2.225 2.225 0 0 0-2.235-2.25Z" />
+                  <path id="Vector_15" d="M387.799 44.634a1.756 1.756 0 1 0 0-3.512 1.756 1.756 0 0 0 0 3.512Z" />
+                  <path id="Vector_16" d="m400.553 50.715-1.165-4.406a1.62 1.62 0 0 0-1.564-1.213c-.735 0-1.389.495-1.565 1.213l-1.165 4.406a.735.735 0 0 0 .718.926h.83l.447 1.82a.562.562 0 0 0 .543.431h.415a.562.562 0 0 0 .543-.43l.431-1.82h.83c.479 0 .83-.464.702-.927Z" />
+                  <path id="Vector_17" d="M397.823 44.634a1.756 1.756 0 1 0 0-3.512 1.756 1.756 0 0 0 0 3.512Z" />
+                  <path id="Rectangle 99_4" d="M392.188 41.108h1.247v12.774h-1.247z" />
+                </g>
+              </g>
+            </g>
+          </g>
+        </g>
+        <g id="showers" {...(isDesktop && {onMouseEnter: () => onMouseEnter && onMouseEnter('showers'), onMouseLeave})}>
+          <rect width="47.1631" height="268" transform="translate(216 0.5)" fill="white" />
+          <g id="Frame 200_2">
+            <rect x="217.498" y="1.99788" width="44.1674" height="44.1674" stroke="#454545" strokeWidth="2.99575" />
+            <path id="Vector_18" fillRule="evenodd" clipRule="evenodd" d="M229.628 31.5836C232.646 31.5836 235.11 29.243 235.11 26.2422C235.11 24.0816 231.168 18.8602 229.751 16.5796C228.273 19.1003 224.146 24.0816 224.146 26.2422C224.146 29.243 226.609 31.5836 229.628 31.5836Z" fill="#454545" />
+            <path id="Vector_19" d="M255.018 28.632C255.018 29.1727 254.866 29.6976 254.584 30.1514C254.302 30.6052 253.885 30.9911 253.359 31.2259C253.01 31.3817 252.615 31.4683 252.199 31.4673H241.929C241.514 31.4683 241.118 31.3817 240.769 31.2248C240.419 31.069 240.116 30.8455 239.871 30.5797C239.379 30.0469 239.111 29.3517 239.11 28.632C239.109 28.2427 239.19 27.8455 239.364 27.4706C239.536 27.0968 239.803 26.7499 240.152 26.4661L245.745 21.9112L245.482 21.6977C245.14 21.4197 244.878 21.0736 244.704 20.6988C244.53 20.3229 244.444 19.918 244.444 19.5142C244.444 19.1104 244.53 18.7056 244.703 18.3297C244.876 17.9538 245.139 17.6066 245.482 17.3276C245.998 16.9072 246.632 16.6958 247.26 16.6958C247.889 16.6958 248.522 16.9082 249.038 17.3286L249.865 18.0038L248.812 19.2976L247.984 18.6224C247.771 18.45 247.518 18.3644 247.26 18.3644C247.001 18.3644 246.746 18.45 246.534 18.6224C246.391 18.7403 246.287 18.8772 246.217 19.0296C246.147 19.181 246.112 19.3479 246.112 19.5145C246.112 19.6803 246.147 19.8459 246.217 19.9972C246.287 20.1486 246.391 20.2865 246.535 20.4045L247.591 21.2644L253.977 26.4661C254.326 26.7499 254.591 27.0968 254.765 27.4706C254.938 27.8455 255.018 28.2427 255.018 28.632ZM253.267 29.07C253.323 28.93 253.351 28.7797 253.351 28.6317C253.351 28.4728 253.318 28.3169 253.252 28.1724C253.184 28.0266 253.083 27.89 252.924 27.7597L247.064 22.9868L241.205 27.7599C241.046 27.89 240.944 28.0268 240.877 28.1727C240.81 28.3172 240.778 28.473 240.778 28.632C240.777 28.8534 240.842 29.0813 240.961 29.2706C241.079 29.462 241.245 29.611 241.448 29.7023C241.584 29.7623 241.739 29.7992 241.929 29.7992H252.199C252.341 29.7992 252.463 29.7781 252.574 29.7423C252.684 29.7068 252.782 29.6555 252.87 29.5923C253.047 29.4651 253.186 29.2803 253.267 29.07Z" fill="#454545" />
+          </g>
+          <g id="Frame 206">
+            <rect x="217.498" y="46.1654" width="44.1674" height="44.1674" stroke="#454545" strokeWidth="2.99575" />
+            <path id="Vector_20" fillRule="evenodd" clipRule="evenodd" d="M229.628 75.7511C232.646 75.7511 235.11 73.4105 235.11 70.4097C235.11 68.2491 231.168 63.0277 229.751 60.7471C228.273 63.2677 224.146 68.2491 224.146 70.4097C224.146 73.4105 226.609 75.7511 229.628 75.7511Z" fill="#454545" />
+            <path id="Vector_21" d="M255.018 72.7995C255.018 73.3402 254.866 73.865 254.584 74.3189C254.302 74.7727 253.885 75.1586 253.359 75.3934C253.01 75.5492 252.615 75.6358 252.199 75.6348H241.929C241.514 75.6358 241.118 75.5492 240.769 75.3923C240.419 75.2365 240.116 75.013 239.871 74.7471C239.379 74.2144 239.111 73.5192 239.11 72.7995C239.109 72.4102 239.19 72.013 239.364 71.6381C239.536 71.2643 239.803 70.9174 240.152 70.6336L245.745 66.0787L245.482 65.8652C245.14 65.5872 244.878 65.2411 244.704 64.8662C244.53 64.4903 244.444 64.0855 244.444 63.6817C244.444 63.2779 244.53 62.873 244.703 62.4972C244.876 62.1213 245.139 61.7741 245.482 61.495C245.998 61.0747 246.632 60.8633 247.26 60.8633C247.889 60.8633 248.522 61.0757 249.038 61.4961L249.865 62.1713L248.812 63.4651L247.984 62.7899C247.771 62.6175 247.518 62.5319 247.26 62.5319C247.001 62.5319 246.746 62.6175 246.534 62.7899C246.391 62.9078 246.287 63.0447 246.217 63.1971C246.147 63.3484 246.112 63.5153 246.112 63.682C246.112 63.8478 246.147 64.0134 246.217 64.1647C246.287 64.3161 246.391 64.454 246.535 64.5719L247.591 65.4319L253.977 70.6336C254.326 70.9174 254.591 71.2643 254.765 71.6381C254.938 72.013 255.018 72.4102 255.018 72.7995ZM253.267 73.2375C253.323 73.0975 253.351 72.9472 253.351 72.7992C253.351 72.6402 253.318 72.4844 253.252 72.3399C253.184 72.1941 253.083 72.0574 252.924 71.9271L247.064 67.1542L241.205 71.9274C241.046 72.0574 240.944 72.1943 240.877 72.3402C240.81 72.4847 240.778 72.6405 240.778 72.7995C240.777 73.0209 240.842 73.2488 240.961 73.4381C241.079 73.6295 241.245 73.7784 241.448 73.8698C241.584 73.9298 241.739 73.9667 241.929 73.9667H252.199C252.341 73.9667 252.463 73.9456 252.574 73.9098C252.684 73.8743 252.782 73.8229 252.87 73.7598C253.047 73.6326 253.186 73.4478 253.267 73.2375Z" fill="#454545" />
+          </g>
+          <g id="Frame 207">
+            <rect x="217.498" y="90.3328" width="44.1674" height="44.1674" stroke="#454545" strokeWidth="2.99575" />
+            <path id="Vector_22" fillRule="evenodd" clipRule="evenodd" d="M229.628 119.919C232.646 119.919 235.11 117.578 235.11 114.577C235.11 112.417 231.168 107.195 229.751 104.915C228.273 107.435 224.146 112.417 224.146 114.577C224.146 117.578 226.609 119.919 229.628 119.919Z" fill="#454545" />
+            <path id="Vector_23" d="M255.018 116.967C255.018 117.508 254.866 118.033 254.584 118.486C254.302 118.94 253.885 119.326 253.359 119.561C253.01 119.717 252.615 119.803 252.199 119.802H241.929C241.514 119.803 241.118 119.717 240.769 119.56C240.419 119.404 240.116 119.18 239.871 118.915C239.379 118.382 239.111 117.687 239.11 116.967C239.109 116.578 239.19 116.18 239.364 115.806C239.536 115.432 239.803 115.085 240.152 114.801L245.745 110.246L245.482 110.033C245.14 109.755 244.878 109.409 244.704 109.034C244.53 108.658 244.444 108.253 244.444 107.849C244.444 107.445 244.53 107.041 244.703 106.665C244.876 106.289 245.139 105.942 245.482 105.663C245.998 105.242 246.632 105.031 247.26 105.031C247.889 105.031 248.522 105.243 249.038 105.664L249.865 106.339L248.812 107.633L247.984 106.957C247.771 106.785 247.518 106.699 247.26 106.699C247.001 106.699 246.746 106.785 246.534 106.957C246.391 107.075 246.287 107.212 246.217 107.365C246.147 107.516 246.112 107.683 246.112 107.849C246.112 108.015 246.147 108.181 246.217 108.332C246.287 108.484 246.391 108.621 246.535 108.739L247.591 109.599L253.977 114.801C254.326 115.085 254.591 115.432 254.765 115.806C254.938 116.18 255.018 116.578 255.018 116.967ZM253.267 117.405C253.323 117.265 253.351 117.115 253.351 116.967C253.351 116.808 253.318 116.652 253.252 116.507C253.184 116.362 253.083 116.225 252.924 116.095L247.064 111.322L241.205 116.095C241.046 116.225 240.944 116.362 240.877 116.508C240.81 116.652 240.778 116.808 240.778 116.967C240.777 117.188 240.842 117.416 240.961 117.606C241.079 117.797 241.245 117.946 241.448 118.037C241.584 118.097 241.739 118.134 241.929 118.134H252.199C252.341 118.134 252.463 118.113 252.574 118.077C252.684 118.042 252.782 117.99 252.87 117.927C253.047 117.8 253.186 117.615 253.267 117.405Z" fill="#454545" />
+          </g>
+          <g id="Frame 208">
+            <rect x="217.498" y="134.5" width="44.1674" height="44.1674" stroke="#454545" strokeWidth="2.99575" />
+            <path id="Vector_24" fillRule="evenodd" clipRule="evenodd" d="M229.628 164.086C232.646 164.086 235.11 161.745 235.11 158.744C235.11 156.584 231.168 151.362 229.751 149.082C228.273 151.602 224.146 156.584 224.146 158.744C224.146 161.745 226.609 164.086 229.628 164.086Z" fill="#454545" />
+            <path id="Vector_25" d="M255.018 161.134C255.018 161.675 254.866 162.2 254.584 162.653C254.302 163.107 253.885 163.493 253.359 163.728C253.01 163.884 252.615 163.97 252.199 163.969H241.929C241.514 163.97 241.118 163.884 240.769 163.727C240.419 163.571 240.116 163.347 239.871 163.082C239.379 162.549 239.111 161.854 239.11 161.134C239.109 160.745 239.19 160.347 239.364 159.973C239.536 159.599 239.803 159.252 240.152 158.968L245.745 154.413L245.482 154.2C245.14 153.922 244.878 153.576 244.704 153.201C244.53 152.825 244.444 152.42 244.444 152.016C244.444 151.612 244.53 151.208 244.703 150.832C244.876 150.456 245.139 150.109 245.482 149.83C245.998 149.409 246.632 149.198 247.26 149.198C247.889 149.198 248.522 149.41 249.038 149.831L249.865 150.506L248.812 151.8L247.984 151.124C247.771 150.952 247.518 150.866 247.26 150.866C247.001 150.866 246.746 150.952 246.534 151.124C246.391 151.242 246.287 151.379 246.217 151.532C246.147 151.683 246.112 151.85 246.112 152.016C246.112 152.182 246.147 152.348 246.217 152.499C246.287 152.651 246.391 152.788 246.535 152.906L247.591 153.766L253.977 158.968C254.326 159.252 254.591 159.599 254.765 159.973C254.938 160.347 255.018 160.745 255.018 161.134ZM253.267 161.572C253.323 161.432 253.351 161.282 253.351 161.134C253.351 160.975 253.318 160.819 253.252 160.674C253.184 160.529 253.083 160.392 252.924 160.262L247.064 155.489L241.205 160.262C241.046 160.392 240.944 160.529 240.877 160.675C240.81 160.819 240.778 160.975 240.778 161.134C240.777 161.355 240.842 161.583 240.961 161.773C241.079 161.964 241.245 162.113 241.448 162.204C241.584 162.264 241.739 162.301 241.929 162.301H252.199C252.341 162.301 252.463 162.28 252.574 162.244C252.684 162.209 252.782 162.157 252.87 162.094C253.047 161.967 253.186 161.782 253.267 161.572Z" fill="#454545" />
+          </g>
+          <g id="Frame 209">
+            <rect x="217.498" y="178.667" width="44.1674" height="44.1674" stroke="#454545" strokeWidth="2.99575" />
+            <path id="Vector_26" fillRule="evenodd" clipRule="evenodd" d="M229.628 208.253C232.646 208.253 235.11 205.912 235.11 202.912C235.11 200.751 231.168 195.53 229.751 193.249C228.273 195.77 224.146 200.751 224.146 202.912C224.146 205.912 226.609 208.253 229.628 208.253Z" fill="#454545" />
+            <path id="Vector_27" d="M255.018 205.301C255.018 205.842 254.866 206.367 254.584 206.821C254.302 207.275 253.885 207.661 253.359 207.895C253.01 208.051 252.615 208.138 252.199 208.137H241.929C241.514 208.138 241.118 208.051 240.769 207.894C240.419 207.738 240.116 207.515 239.871 207.249C239.379 206.716 239.111 206.021 239.11 205.301C239.109 204.912 239.19 204.515 239.364 204.14C239.536 203.766 239.803 203.419 240.152 203.136L245.745 198.581L245.482 198.367C245.14 198.089 244.878 197.743 244.704 197.368C244.53 196.992 244.444 196.587 244.444 196.184C244.444 195.78 244.53 195.375 244.703 194.999C244.876 194.623 245.139 194.276 245.482 193.997C245.998 193.577 246.632 193.365 247.26 193.365C247.889 193.365 248.522 193.578 249.038 193.998L249.865 194.673L248.812 195.967L247.984 195.292C247.771 195.119 247.518 195.034 247.26 195.034C247.001 195.034 246.746 195.119 246.534 195.292C246.391 195.41 246.287 195.547 246.217 195.699C246.147 195.85 246.112 196.017 246.112 196.184C246.112 196.35 246.147 196.515 246.217 196.667C246.287 196.818 246.391 196.956 246.535 197.074L247.591 197.934L253.977 203.136C254.326 203.419 254.591 203.766 254.765 204.14C254.938 204.515 255.018 204.912 255.018 205.301ZM253.267 205.739C253.323 205.599 253.351 205.449 253.351 205.301C253.351 205.142 253.318 204.986 253.252 204.842C253.184 204.696 253.083 204.559 252.924 204.429L247.064 199.656L241.205 204.429C241.046 204.559 240.944 204.696 240.877 204.842C240.81 204.987 240.778 205.142 240.778 205.301C240.777 205.523 240.842 205.751 240.961 205.94C241.079 206.131 241.245 206.28 241.448 206.372C241.584 206.432 241.739 206.469 241.929 206.469H252.199C252.341 206.469 252.463 206.448 252.574 206.412C252.684 206.376 252.782 206.325 252.87 206.262C253.047 206.135 253.186 205.95 253.267 205.739Z" fill="#454545" />
+          </g>
+          <g id="Frame 210">
+            <rect x="217.498" y="222.835" width="44.1674" height="44.1674" stroke="#454545" strokeWidth="2.99575" />
+            <path id="Vector_28" fillRule="evenodd" clipRule="evenodd" d="M229.628 252.421C232.646 252.421 235.11 250.08 235.11 247.079C235.11 244.919 231.168 239.697 229.751 237.417C228.273 239.937 224.146 244.919 224.146 247.079C224.146 250.08 226.609 252.421 229.628 252.421Z" fill="#454545" />
+            <path id="Vector_29" d="M255.018 249.469C255.018 250.01 254.866 250.534 254.584 250.988C254.302 251.442 253.885 251.828 253.359 252.063C253.01 252.219 252.615 252.305 252.199 252.304H241.929C241.514 252.305 241.118 252.219 240.769 252.062C240.419 251.906 240.116 251.682 239.871 251.417C239.379 250.884 239.111 250.189 239.11 249.469C239.109 249.08 239.19 248.682 239.364 248.308C239.536 247.934 239.803 247.587 240.152 247.303L245.745 242.748L245.482 242.535C245.14 242.257 244.878 241.911 244.704 241.536C244.53 241.16 244.444 240.755 244.444 240.351C244.444 239.947 244.53 239.542 244.703 239.167C244.876 238.791 245.139 238.443 245.482 238.164C245.998 237.744 246.632 237.533 247.26 237.533C247.889 237.533 248.522 237.745 249.038 238.166L249.865 238.841L248.812 240.134L247.984 239.459C247.771 239.287 247.518 239.201 247.26 239.201C247.001 239.201 246.746 239.287 246.534 239.459C246.391 239.577 246.287 239.714 246.217 239.867C246.147 240.018 246.112 240.185 246.112 240.351C246.112 240.517 246.147 240.683 246.217 240.834C246.287 240.986 246.391 241.123 246.535 241.241L247.591 242.101L253.977 247.303C254.326 247.587 254.591 247.934 254.765 248.308C254.938 248.682 255.018 249.08 255.018 249.469ZM253.267 249.907C253.323 249.767 253.351 249.617 253.351 249.469C253.351 249.31 253.318 249.154 253.252 249.009C253.184 248.863 253.083 248.727 252.924 248.597L247.064 243.824L241.205 248.597C241.046 248.727 240.944 248.864 240.877 249.01C240.81 249.154 240.778 249.31 240.778 249.469C240.777 249.69 240.842 249.918 240.961 250.108C241.079 250.299 241.245 250.448 241.448 250.539C241.584 250.599 241.739 250.636 241.929 250.636H252.199C252.341 250.636 252.463 250.615 252.574 250.579C252.684 250.544 252.782 250.492 252.87 250.429C253.047 250.302 253.186 250.117 253.267 249.907Z" fill="#454545" />
+          </g>
+        </g>
+      </svg>
+    </>
+  )
+}
